Recurse through inherited as well as non-inh groups
Previously check_access() declined to recurse into group roles when
the current role had the inherit attribute, the theory being that it
was unnecessary since all of those privileges were already inherited.
However this fails to consider the situation where there is a chain
of granted roles, and somewhere further up the chain there is a role
with noinherit. In this case the reported privileges for the base
role do not reflect the non-inherited privileges. Fix that by
recursing always, and make the chain of grants clearer by reflecting
at each level whether the role is inherit -- (true) -- or noinherit
-- (false).

Additionally, while at it, fix .gitignore. The check_access.sql file
was being ignored, which is unfortunate since that is where all of
the substance of the extension exists.